should fix css and import issues addressed in #26
