+++ package.json
@@ -21,7 +21,7 @@
     "react-app-rewired": "^2.1.6",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
+    "react-scripts": "3.3.0",
     "react-toastify": "^6.0.8",
     "socket.io-client": "^2.3.0"
   },
