+++ src/shared/components/loaders/Spinner/index.js
@@ -1,4 +1,3 @@
-
 import Spinner from "./Spinner";
