+++ src/shared/context/auth/index.js
@@ -1,10 +1,12 @@
 import withAuthProvider from "./withAuthProvider";
 import withAuthCheck from "./withAuthCheck";
+import withAuthPage from "./withAuthPage";
 import AuthContext from "./AuthContext";
 import useAuth from "./useAuth";
 import AuthProvider from "./AuthProvider";
 
 export {
+    withAuthPage,
     withAuthCheck,
     withAuthProvider,
     AuthContext,
